Add plot function; initialize with random uniform
An accuracy plot function is added which shows the accuracy for both
training and validation data.
The weights and biases are initialized with random uniform values
instead of random normal values, but this does not show great effect to
the moment.
A minor bug concerning the learning_rate is fixed.